Move linearise() to GraphPlanner
