+++ traversal/procedure/GraphProcedure.java
@@ -44,11 +44,9 @@ import org.slf4j.LoggerFactory;
 import javax.annotation.Nullable;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import static com.vaticle.typedb.core.common.iterator.Iterators.iterate;
 import static com.vaticle.typedb.core.concurrent.producer.Producers.async;
@@ -185,27 +183,8 @@ public class GraphProcedure implements PermutationProcedure {
         }
 
         public GraphProcedure build() {
-            linearise();
             return new GraphProcedure(vertices.values().stream().sorted(comparing(ProcedureVertex::order))
                     .toArray(ProcedureVertex[]::new));
-        }
-
-        private void linearise() {
-            Set<ProcedureVertex<?, ?>> orderedSet = new HashSet<>();
-            List<ProcedureVertex<?, ?>> stack =  vertices.values().stream().filter(ProcedureVertex::isStartVertex).collect(Collectors.toList());
-            int vertexOrder = 0;
-            while (!stack.isEmpty()) {
-                ProcedureVertex<?, ?> vertex = stack.remove(stack.size() - 1);
-                vertex.setOrder(vertexOrder++);
-                orderedSet.add(vertex);
-                for (ProcedureVertex<?, ?> v : vertex.outs().stream().map(ProcedureEdge::to).collect(Collectors.toSet())) {
-                    if (!orderedSet.contains(v) && orderedSet.containsAll(v.ins().stream().map(ProcedureEdge::from).collect(Collectors.toSet()))) {
-                        stack.add(v);
-                    }
-                }
-            }
-            assert orderedSet.size() == vertices.size();
-            assert vertexOrder == vertices.size();
         }
 
         private void register(GraphPlanner planner, int startOrder) {
